Update ComunityGoalManager.cs and NarrativeProgressBarObject.cs

+++ Assets/ComunityGoalManager.cs
@@ -56,7 +56,7 @@ public class ComunityGoalManager : MonoBehaviour
         }
 
         if (personalContribution != null) {
-            personalContribution.text = "Personal Contribution: " + value.DoneYou;
+            personalContribution.text = "Personal Contribution: " + (int)value.DoneYou;
         }
         UnityEngine.UI.LayoutRebuilder.ForceRebuildLayoutImmediate(GetComponent<RectTransform>());
     }
